refactors sidebar to include a sign out button; implements store.signOut()
