+++ src/routes/signup/__layout.reset.svelte
@@ -1,8 +1,9 @@
 <script lang="ts">
     import { setContext } from "svelte";
     import Layout from "$lib/layouts/onboarding/Layout.svelte";
-    import { store } from "$lib/stores/app-store";
+    import { store, isAuthenticated } from "$lib/stores/app-store";
     setContext("rally:store", store);
+    setContext("rally:isAuthenticated", isAuthenticated);
 </script>
 
 <svelte:head>
