+++ src/routes/terms/index.svelte
@@ -1,14 +1,16 @@
 <script lang="ts">
   import { getContext } from "svelte";
-  import { goto } from "$app/navigation";
   import TermsContent from "$lib/views/terms/Content.svelte";
+  import redirectFromMainViewIfNeeded from "../_redirect-from-main-view-if-needed";
 
   import type { Readable } from "svelte/store";
+  import type { AppStore } from "$lib/stores/types";
+
+  const store: AppStore = getContext("rally:store");
   const isAuthenticated :Readable<boolean> = getContext("rally:isAuthenticated");
 
-  $: if ($isAuthenticated === false) {
-    goto("/signup");
-  }
+  $: redirectFromMainViewIfNeeded($isAuthenticated, $store._initialized, !$store?.user?.enrolled);
+
 </script>
 
 <svelte:head>
@@ -17,4 +19,4 @@
 
 {#if $isAuthenticated}
   <TermsContent />
-{/if}
+{/if}3
